tests: moved test over
