+++ tests/tests_unit/rules/test_models/test_entities.py
@@ -1,10 +1,13 @@
 from typing import Any
 
 import pytest
+from cognite.client.data_classes.data_modeling import DataModelId, PropertyId, ViewId
 
 from cognite.neat.rules.models.entities import (
     ClassEntity,
+    DataModelEntity,
     Entity,
+    ReferenceEntity,
     Undefined,
     Unknown,
     ViewEntity,
@@ -59,3 +62,82 @@ class TestEntities:
         loaded = cls_.load(raw, space=DEFAULT_SPACE, version=DEFAULT_VERSION)
 
         assert loaded == expected
+
+
+class TestDataModelEntity:
+    @pytest.mark.parametrize(
+        "raw, expected, expected_id",
+        [
+            (
+                "test:TestGraphQL1(version=3)",
+                DataModelEntity(space="test", externalId="TestGraphQL1", version="3"),
+                DataModelId("test", "TestGraphQL1", "3"),
+            ),
+            (
+                "test:TestGraphQL1",
+                DataModelEntity(space="test", externalId="TestGraphQL1", version=DEFAULT_VERSION),
+                DataModelId("test", "TestGraphQL1", DEFAULT_VERSION),
+            ),
+        ],
+    )
+    def test_load(self, raw: str, expected: DataModelEntity, expected_id: DataModelId) -> None:
+        actual = DataModelEntity.load(raw, space=DEFAULT_SPACE, version=DEFAULT_VERSION)
+
+        assert actual == expected
+        assert actual.as_id() == expected_id
+
+
+class TestViewPropertyEntity:
+    @pytest.mark.parametrize(
+        "raw, expected, expected_id",
+        [
+            pytest.param(
+                "test:TestGraphQL1(version=3, property=prop1)",
+                ViewPropertyEntity(space="test", externalId="TestGraphQL1", version="3", property="prop1"),
+                PropertyId(ViewId("test", "TestGraphQL1", "3"), "prop1"),
+                id="Prefix, suffix, version and prop",
+            ),
+            pytest.param(
+                "test:TestGraphQL1(property=prop1)",
+                ViewPropertyEntity(space="test", externalId="TestGraphQL1", version=DEFAULT_VERSION, property="prop1"),
+                PropertyId(ViewId("test", "TestGraphQL1", DEFAULT_VERSION), "prop1"),
+                id="Prefix, suffix and prop. Skip version",
+            ),
+        ],
+    )
+    def test_load(self, raw: str, expected: ViewPropertyEntity, expected_id: PropertyId) -> None:
+        actual = ViewPropertyEntity.load(raw, space=DEFAULT_SPACE, version=DEFAULT_VERSION)
+
+        assert actual == expected
+        assert actual.as_id() == expected_id
+
+
+class TestReferenceType:
+    @pytest.mark.parametrize(
+        "raw, expected,",
+        [
+            pytest.param(
+                "test:TestGraphQL1(version=3, property=prop1)",
+                ReferenceEntity(prefix="test", suffix="TestGraphQL1", version="3", property="prop1"),
+                id="Prefix, suffix, version and prop",
+            ),
+            pytest.param(
+                "test:TestGraphQL1(property=prop1)",
+                ReferenceEntity(prefix="test", suffix="TestGraphQL1", version=None, property="prop1"),
+                id="Prefix, suffix and prop. Skip version",
+            ),
+            pytest.param(
+                "test:TestGraphQL1(property=test:prop1)",
+                ReferenceEntity(prefix="test", suffix="TestGraphQL1", version=None, property="test:prop1"),
+                id="Prefix, suffix and prop. Skip version",
+            ),
+            pytest.param(
+                "test:TestGraphQL1",
+                ReferenceEntity(prefix="test", suffix="TestGraphQL1", version=None, property=None),
+                id="Prefix, suffix and prop. Skip version",
+            ),
+        ],
+    )
+    def test_load(self, raw: str, expected: ReferenceEntity) -> None:
+        actual = ReferenceEntity.load(raw)
+        assert actual == expected
